backend/requirements.txt
Flask==3.1.3
python-dotenv>=1.0.0
cdsapi>=0.7.4
xarray>=2024.1.0
netCDF4>=1.7.0
numpy>=1.26.0
